Changing float to int

+++ RSound-A/Program.cs
@@ -24,20 +24,20 @@ namespace Encryption
 {
     public class RSA
     {
-        public static float privateKey;
-        public static List<float> publickey;
+        public static int privateKey;
+        public static List<int> publickey;
 
         public RSA()
         {
            // GenerateKey();
         }
 
-        public float Crypt(float toCrypt)
+        public int Crypt(int toCrypt)
         {
             return PowerModulo(toCrypt, publickey[1], publickey[0]);
         }
 
-        public float Decrypt(float toDecrypt)
+        public int Decrypt(int toDecrypt)
         {
             return PowerModulo(toDecrypt, privateKey, publickey[0]);
         }
@@ -53,11 +53,11 @@ namespace Encryption
 
         }
 
-        public static float PowerModulo(float M, float e, float mod)
+        public static int PowerModulo(int M, int e, int mod)
         {
             //Calcul de c = M^e [mod]
-            float c = M;
-            for (float a = 1; a < e; ++a) 
+            int c = M;
+            for (int a = 1; a < e; ++a) 
                 c = (M * c) % mod;
             return c;
         }
@@ -69,7 +69,7 @@ namespace Encryption
 public class RSAKey
 {
 
-    private static List<float> PrimaryNumbers = new List<float>{
+    private static List<int> PrimaryNumbers = new List<int>{
             7009421, 7009427, 7009447, 7009469, 7009477, 7009481,
             7009487, 7009493, 7009507, 7009517, 7009543, 7009547, 7009553, 7009577, 7009601, 7009603, 7009609,
             7009663, 7009669, 7009699, 7009711, 7009757, 7009771, 7009777, 7009799, 7009801, 7009823, 7009859,
@@ -83,11 +83,11 @@ public class RSAKey
         };
 
 
-    private static float Q;
-    private static float P;
-    private static float N;
-    private static float E;
-    private static float Phi;
+    private static int Q;
+    private static int P;
+    private static int N;
+    private static int E;
+    private static int Phi;
 
     public RSAKey()
     {
@@ -102,28 +102,28 @@ public class RSAKey
 
      }
 
-    private static float GetRandomPrimaryNumber()
+    private static int GetRandomPrimaryNumber()
     {
         return PrimaryNumbers[new Random().Next(0, PrimaryNumbers.Count)];
     }
 
  
 
-    private static float PGCD(float a, float b)
+    private static int PGCD(int a, int b)
     {
-         float modulo = a % b;
+         int modulo = a % b;
         if (modulo == 0) return b;
         return PGCD(b, modulo);
         
        // return (a%b==0) ? b : (PGCD(b, a%b));
     }
 
-    private static float GetEncryptionExponent()
+    private static int GetEncryptionExponent()
     {
 
         Random rnd = new Random();
         while (true) {
-            float e = rnd.Next(0, 100);
+            int e = rnd.Next(0, 100);
              //MessageBox.Show(Phi.ToString());
 
             if ((PGCD(e, ((P - 1) * (Q - 1))) == 1) && (e < Phi)) {
@@ -135,20 +135,20 @@ public class RSAKey
 
     }
 
-    public List<float> GetPublicKey()
+    public List<int> GetPublicKey()
     {
 
-       return new List<float> { N, E };
+       return new List<int> { N, E };
     }
 
-    public float GetPrivateKey()
+    public int GetPrivateKey()
     {
         return GetDecryptionExponent();
     }
 
 
 
-    private static float GetDecryptionExponent()
+    private static int GetDecryptionExponent()
     {
         //calculer l'entier naturel d, inverse de e modulo φ(n), et strictement inférieur à φ(n), 
         //appelé exposant de déchiffrement ; d peut se calculer efficacement par l'algorithme d'Euclide étendu.
@@ -159,12 +159,12 @@ public class RSAKey
     }
 
 
-    private static float modInverse(float a, float n)
+    private static int modInverse(int a, int n)
     {
         //Marche pas
-        float i = n, v = 0, d = 1;
+        int i = n, v = 0, d = 1;
         while (a > 0) {
-            float t = i / a, x = a;
+            int t = i / a, x = a;
             a = i % x;
             i = x;
             x = d;
